Add files via upload
The CataractDetector is a Python-based software application designed to detect the presence and severity of cataracts in eye images. Utilizing computer vision techniques, including the Haar Cascade Classifier for eye detection and the Hough Circle Transform for identifying the lens, the application provides an automated method for evaluating cataracts. The system preprocesses images, detects eyes, identifies circular patterns indicative of cataracts, and analyzes the average color intensity within the eye region to determine the presence and extent of cataracts. The outcome is visually marked on the image and provides a quantifiable assessment of cataract severity